Move all init functions into regions/init.cc.
